len always returns int

+++ Python/004-largest_palindrome_product.py
@@ -11,7 +11,7 @@ Find the largest palindrome made from the product of two 3-digit numbers.
 def is_palindrome(string):
     """ Brute force algorithm """
     i, palindrome = 0, True
-    while i<=int(len(string)/2) and palindrome == True:
+    while i<=len(string)/2 and palindrome == True:
         if string[i] != string[-(i+1)]:
             palindrome = False
         i+=1
